New question, SOCVR Code of Conduct
This is related to our new pun-master RO answer [here](https://gist.github.com/mogsdad/c22cbb03cb7de02838cb#gistcomment-1721488):

> WE ARE ALWAYS AMBASSADORS FOR THE ROOM

This is totally badass and I think that it'd be great to have something more than "Be Nice" to offer our members. I'd like to discuss the idea of the redaction of a Code of Conduct during the meeting.  
The [General expectations for members](https://github.com/SO-Close-Vote-Reviewers/SOCVR-RoomInformation#general-expectations-for-members) is just a bunch of specific guidelines we've aggregated over the years. It does not cover the broader picture. The last rules are interesting but not developed enough.